feat: implement file manager page for desktop file transfers

+++ flutter/lib/desktop/pages/file_manager_page.dart
@@ -1371,7 +1371,7 @@ class _FileManagerViewState extends State<FileManagerView> {
   bool _checkDoubleClick(Entry entry) {
     final current = DateTime.now().millisecondsSinceEpoch;
     final elapsed = current - _lastClickTime;
-    _lastClickTime = current.toDouble();
+    _lastClickTime = current;
     if (_lastClickEntry == entry) {
       if (elapsed < bind.getDoubleClickTime()) {
         return true;
